Port PNNS tests to _TestUtilities
