+++ Sources/PrivateNearestNeighborSearch/MatrixMultiplication.swift
@@ -1,4 +1,4 @@
-// Copyright 2024 Apple Inc. and the Swift Homomorphic Encryption project authors
+// Copyright 2024-2025 Apple Inc. and the Swift Homomorphic Encryption project authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -117,7 +117,7 @@ extension PlaintextMatrix {
     /// - Returns: Encrypted dense-column packed vector containing dot products.
     /// - Throws: Error upon failure to compute the inner product.
     @inlinable
-    func mulTranspose(
+    package func mulTranspose(
         vector ciphertextVector: CiphertextMatrix<Scheme, Scheme.CanonicalCiphertextFormat>,
         using evaluationKey: EvaluationKey<Scheme>) throws -> [Scheme.CanonicalCiphertext]
     {
@@ -215,7 +215,7 @@ extension PlaintextMatrix {
     /// - Returns: Encrypted dense-column packed matrix.
     /// - Throws: Error upon failure to compute the product.
     @inlinable
-    func mulTranspose(
+    package func mulTranspose(
         matrix ciphertextMatrix: CiphertextMatrix<Scheme, Scheme.CanonicalCiphertextFormat>,
         using evaluationKey: EvaluationKey<Scheme>) throws
         -> CiphertextMatrix<Scheme, Scheme.CanonicalCiphertextFormat>
